model/requirements.txt
numpy>=1.24.0
matplotlib>=3.7.0
pyserial>=3.5